refactor(db): Consolidate and seed database migrations
