--- alembic/versions/66258ea5da46_initial_database_schema_creation.py
@@ -1,32 +1,0 @@
-"""Initial database schema creation
-
-Revision ID: 66258ea5da46
-Revises: 6a3a65746d29
-Create Date: 2025-11-06 21:10:59.469104
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '66258ea5da46'
-down_revision: Union[str, Sequence[str], None] = '6a3a65746d29'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
