+++ alembic/versions/6a3a65746d29_initial_database_schema_creation.py
@@ -13,7 +13,7 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision: str = '6a3a65746d29'
-down_revision: Union[str, Sequence[str], None] = '356c75e8ffb6'
+down_revision: Union[str, Sequence[str], None] = None
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
