+++ alembic/versions/f8f6b4b4a9a9_setup_initial_data.py
@@ -1,0 +1,56 @@
+"""setup_initial_data
+
+Revision ID: f8f6b4b4a9a9
+Revises: 6a3a65746d29
+Create Date: 2025-11-20 12:00:00.000000
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'f8f6b4b4a9a9'
+down_revision: Union[str, Sequence[str], None] = '6a3a65746d29'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    transaction_types_table = sa.table(
+        'transaction_types',
+        sa.column('name', sa.Text),
+        sa.column('icon', sa.Text),
+    )
+    op.bulk_insert(
+        transaction_types_table,
+        [
+            {'name': 'expense', 'icon': '💸'},
+            {'name': 'income', 'icon': '💰'},
+            {'name': 'transfer', 'icon': '➡️'},
+        ]
+    )
+
+    default_categories_table = sa.table(
+        'default_categories',
+        sa.column('name', sa.Text),
+    )
+    op.bulk_insert(
+        default_categories_table,
+        [
+            {'name': 'Їжа'},
+            {'name': 'Транспорт'},
+            {'name': 'Розваги'},
+        ]
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.execute("DELETE FROM transaction_types WHERE name IN ('expense', 'income', 'transfer')")
+    op.execute("DELETE FROM default_categories WHERE name IN ('Їжа', 'Транспорт', 'Розваги')")
+    # ### end Alembic commands ###
