+++ main.py
@@ -7,9 +7,6 @@ from aiogram.enums import ParseMode
 from aiogram.client.default import DefaultBotProperties
 from src.config.config import BOT_TOKEN
 from src.handlers import router 
-from src.database.manager import setup_initial_data
-from src.database.migrations import run_migrations
-from sqlalchemy.exc import SQLAlchemyError
 
 # Configure logging once at the very beginning
 logging.basicConfig(
@@ -19,43 +16,8 @@ logging.basicConfig(
 )
 logger = logging.getLogger(__name__)
 
-# Maximum number of retries for database connection
-MAX_RETRIES = 10
-RETRY_DELAY_SECONDS = 2 
-
 dp = Dispatcher()
 dp.include_router(router)
-
-async def initialize_database_with_retries():
-    """
-    Asynchronously initializes the database with retry logic and exponential backoff.
-    Raises an exception if the maximum number of retries is reached.
-    """
-    for attempt in range(1, MAX_RETRIES + 1):
-        try:
-            logger.info(f"Attempt {attempt}/{MAX_RETRIES}: launching migrations Alembic...")
-            # Run database migrations
-            await run_migrations()
-            logger.info("✅ migrations completed successfully.")
-            # Adding initial data
-            await setup_initial_data()
-            logger.info("✅ database successfully initialized.")
-            return # Successfully connected and initialized
-        except SQLAlchemyError as e:
-            # Connecting to the migrations failed
-            logger.error(f"❌ Attempt {attempt}/{MAX_RETRIES}: Attempt connection to migrations failed. Error: {e.__class__.__name__} - {e}")
-
-            # If we've reached the max attempts, re-raise the exception
-            if attempt == MAX_RETRIES:
-                logger.critical("🚨 Maximum attempts to connect to migrations reached. Exiting")
-                raise # Re-raise the original exception
-            
-            # To count the wait time before the next retry (exponential backoff, max 30 seconds)
-            wait_time = min(RETRY_DELAY_SECONDS * (2 ** (attempt - 1)), 30)
-                
-            logger.info(f"⏳ wait {wait_time} seconds before the next attempt...")
-            # Use asyncio.sleep for non-blocking wait
-            await asyncio.sleep(wait_time)
 
 # Main asynchronous function to start the bot
 async def main():
@@ -63,14 +25,6 @@ async def main():
     Main entry point for the bot application.
     Initializes the database and starts the bot polling.
     """
-    try:
-        # Initialization of Database with retries
-        # await initialize_database_with_retries()
-        setup_initial_data()
-    except Exception as e: # Catching a general exception if DB init fails
-        logger.error("❌ Failed to initialize database, exiting.", exc_info=e)
-        sys.exit(1) # Important: stop the bot if the DB is not ready
-
     # Initialize the Bot
     bot = Bot(
         token=BOT_TOKEN,
