+++ src/database/manager.py
@@ -185,45 +185,3 @@ if not DATABASE_URL:
 # Initialization happens here, при імпорті модуля
 db_manager = DatabaseManager(DATABASE_URL)
 
-def add_initial_transaction_types():
-    """
-    Ensures that default transaction types (expense, income, transfer) exist in the database.
-    This should be called once after database initialization.
-    """
-    with db_manager.get_session() as session:
-        # Оптимізована перевірка для уникнення кількох запитів до БД
-        expected_types = {'expense', 'income', 'transfer'}
-        existing_types = {tt.name for tt in session.query(TransactionType).filter(TransactionType.name.in_(list(expected_types))).all()}
-
-        for name in expected_types:
-            if name not in existing_types:
-                icon = ''
-                if name == 'expense': icon = '💸'
-                elif name == 'income': icon = '💰'
-                elif name == 'transfer': icon = '➡️'
-                session.add(TransactionType(name=name, icon=icon))
-        logger.info("Initial transaction types added/checked.")
-
-def add_default_categories():
-    """
-    Ensures that a set of predefined default categories exist in the database.
-    These can be used as templates for new users' categories.
-    This should be called once after database initialization.
-    """
-    with db_manager.get_session() as session:
-        default_category_names = ['Їжа', 'Транспорт', 'Розваги']
-        existing_categories = {dc.name for dc in session.query(DefaultCategory).filter(DefaultCategory.name.in_(default_category_names)).all()}
-
-        for name in default_category_names:
-            if name not in existing_categories:
-                session.add(DefaultCategory(name=name))
-        logger.info("Initial default categories added/checked.")
-
-def setup_initial_data():
-    """
-    A convenience function to run all initial data setup tasks.
-    """
-    logger.info("⏳ Setting up initial data (transaction types, default categories)...")
-    add_initial_transaction_types()
-    add_default_categories()
-    logger.info("✅ Initial data setup complete.")
